Version 6
Completed most of the homepage with only the social media links needing to be added to the footer. I also fixed the issue with the navbar expanding every time the dropdown was used. I also began adding the heading and the navbar for the entree page and added the relevant CSS as well.